Dev: unix launch task at the moment

+++ preLaunchTask.sh
@@ -4,28 +4,21 @@
 #
 clear
 
-if [ "$1" = "--webapi" ]; then 
-    
-    kill $(lsof -t -i :3302)
-    if [ $? != 0 ]; then echo "Rescue Shelter webapi//localhost:3302 not open"
-    else echo "Rescue Shelter webapi//localhost:3302 closed"
-    fi
 
+if [ "$1" = "--webapi" ]; then 
+
+    npm run clean
+    npm run compile
+
+    echo "Configure the mongodb docker container"      
+    docDbName=$(sudo docker ps --filter "name=mongo_dev" --format "{{.Names}}")
+    echo "Docker container name: $docDbName"
+    if [ "$docDbName" != "mongo_dev" ]; then
+        sudo docker rm mongo_dev -f
+        sudo docker run -d --rm -p 27017:27017 --name mongo_dev mongo
+    fi
 fi
 
-npm run clean
-npm run compile
-
-echo "Configure the mongodb docker container"
-sudo docker rm mongo_dev -f
-sudo docker run -d --rm -p 27017:27017 --name mongo_dev mongo
-
-if [ "$1" = "--webapp" ]; then 
-    
-    kill $(lsof -t -i :3301)
-    if [ $? != 0 ]; then echo "Rescue Shelter webapp//localhost:3301 not open"
-    else echo "Rescue Shelter webapp//localhost:3301 closed"
-    fi
-
+if [ "$1" = "--webapp" ]; then     
     npm run start 
 fi
